:recycle: 테스트 클래스 공통 코드 추출 (#61)
