+++ ahachul_backend/src/test/kotlin/backend/team/ahachul_backend/config/controller/CommonDocsConfig.kt
@@ -1,0 +1,45 @@
+package backend.team.ahachul_backend.config.controller
+
+import backend.team.ahachul_backend.common.interceptor.AuthenticationInterceptor
+import com.fasterxml.jackson.databind.ObjectMapper
+import org.junit.jupiter.api.BeforeEach
+import org.mockito.BDDMockito
+import org.mockito.Mockito
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs
+import org.springframework.boot.test.mock.mockito.MockBean
+import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext
+import org.springframework.restdocs.operation.preprocess.OperationRequestPreprocessor
+import org.springframework.restdocs.operation.preprocess.OperationResponsePreprocessor
+import org.springframework.restdocs.operation.preprocess.Preprocessors
+import org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessRequest
+import org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse
+import org.springframework.test.web.servlet.MockMvc
+
+@AutoConfigureRestDocs
+abstract class CommonDocsConfig {
+
+    @Autowired lateinit var mockMvc: MockMvc
+    @Autowired lateinit var objectMapper: ObjectMapper
+
+    @MockBean lateinit var authenticationInterceptor: AuthenticationInterceptor
+    @MockBean lateinit var jpaMetamodelMappingContext: JpaMetamodelMappingContext
+
+    @BeforeEach
+    fun setup() {
+        BDDMockito.given(authenticationInterceptor.preHandle(any(), any(), any())).willReturn(true)
+    }
+
+    protected fun getDocsRequest(): OperationRequestPreprocessor {
+        return preprocessRequest(Preprocessors.prettyPrint())
+    }
+
+    protected fun getDocsResponse(): OperationResponsePreprocessor {
+        return preprocessResponse(Preprocessors.prettyPrint())
+    }
+
+    protected fun <T> any(): T {
+        Mockito.any<T>()
+        return null as T
+    }
+}
